requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-bootstrap4==0.0.6
django-sslserver==0.20
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
PyJWT==1.6.4
python-slugify==1.2.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Unidecode==1.0.22
urllib3==1.23
